codex: reuse history normalization for tool search (#30618)
