+++ codex-rs/core/src/context_manager/mod.rs
@@ -5,3 +5,4 @@ pub(crate) mod updates;
 pub(crate) use history::ContextManager;
 pub(crate) use history::is_user_turn_boundary;
 pub(crate) use history::truncate_function_output_payload;
+pub(crate) use normalize::canonicalize_for_durable_history;
